fix(emacs): Disable thin diff bars
:ui vc-gutter +pretty looks bad with themes that invert the
fg/bg of diff-hl’s faces (like modus-themes does).

Doom Emacs uses a soft fork of 'git-gutter-fringe' which introduces
a negative effect on themes that are not designed specifically for Doom:
What once was a plus sign now it turns into a straight line adjacent to
the standard color-coded bg.

https://github.com/doomemacs/doomemacs/blob/master/modules/ui/vc-gutter/README.org
https://protesilaos.com/codelog/2022-08-04-doom-git-gutter-modus-themes/